requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4
    # via pooch
apscheduler==3.6.3
    # via python-telegram-bot
audioread==2.1.9
    # via librosa
backcall==0.2.0
    # via ipython
backports.zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
cachetools==4.2.2
    # via python-telegram-bot
certifi==2021.10.8
    # via
    #   python-telegram-bot
    #   requests
cffi==1.15.0
    # via soundfile
charset-normalizer==2.0.12
    # via requests
colorama==0.4.4
    # via ipython
decorator==5.1.1
    # via
    #   ipython
    #   librosa
    #   traitlets
idna==3.3
    # via requests
importlib-resources==5.4.0
    # via backports.zoneinfo
ipython==7.16.3
    # via -r requirements.in
ipython-genutils==0.2.0
    # via traitlets
jedi==0.17.2
    # via ipython
joblib==1.1.0
    # via
    #   librosa
    #   scikit-learn
librosa==0.9.1
    # via -r requirements.in
llvmlite==0.36.0
    # via numba
numba==0.53.1
    # via
    #   librosa
    #   resampy
numpy==1.19.5
    # via
    #   librosa
    #   numba
    #   resampy
    #   scikit-learn
    #   scipy
packaging==21.3
    # via
    #   librosa
    #   pooch
parso==0.7.1
    # via jedi
pickleshare==0.7.5
    # via ipython
pooch==1.6.0
    # via librosa
prompt-toolkit==3.0.29
    # via ipython
pyaudio==0.2.11
    # via -r requirements.in
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via ipython
pyparsing==3.0.8
    # via packaging
python-telegram-bot==13.11
    # via -r requirements.in
pytz==2022.1
    # via
    #   apscheduler
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.27.1
    # via pooch
resampy==0.2.2
    # via librosa
scikit-learn==0.24.2
    # via librosa
scipy==1.5.4
    # via
    #   librosa
    #   resampy
    #   scikit-learn
six==1.16.0
    # via
    #   apscheduler
    #   resampy
    #   traitlets
soundfile==0.10.3.post1
    # via
    #   -r requirements.in
    #   librosa
speechrecognition==3.8.1
    # via -r requirements.in
threadpoolctl==3.1.0
    # via scikit-learn
tornado==6.1
    # via python-telegram-bot
traitlets==4.3.3
    # via ipython
tzdata==2022.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
tzlocal==4.2
    # via apscheduler
urllib3==1.26.9
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
zipp==3.6.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
